BackEnd/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Cors
Flask-Migrate
Authlib
requests
pycryptodome
psycopg2-binary
python-dotenv
